Rework of deltashare to use default hash set diff
Why:
The existing hashset diff hid the changes needed for comments and the
like. This makes it visible

How:
By using default `SerializeValueForHash` provided by terraform we are
able to diff all the properties that change -- except the properties
that are not in the terraform state. The delta share provider today
use the result from the api fetch as the base for the diff which makes
the diff inaccurate.

For instance, if a table is delta shared outside of terraform, the
diff will not show the removal of that table if it isn't in the
terraform state.

We can see in `TestUpdateShareComplexDiff`. In this test, the state of
the share at the databricks api starts with sharing of table `a`. The
Terraform State at the start of the test *has empty shares*. Then the
HCL adds both table `a` and table `b`. If the terraform provider just
used the Terraform State file, it would make an api call to share both
table `a` and table `b`. This would fail ? because table `a` is
already shared. Therefore, it is best to query the Real State in
Databricks Workspace to determine to only add one table.